Move to different module to avoid class cycle
