+++ subprojects/launcher/src/main/java/org/gradle/launcher/cli/BuildActionsFactory.java
@@ -27,7 +27,7 @@ import org.gradle.internal.concurrent.CompositeStoppable;
 import org.gradle.internal.concurrent.Stoppable;
 import org.gradle.internal.jvm.inspection.JvmVersionDetector;
 import org.gradle.internal.logging.events.OutputEventListener;
-import org.gradle.internal.nativeintegration.ConsoleState;
+import org.gradle.internal.logging.sink.ConsoleState;
 import org.gradle.internal.nativeintegration.services.NativeServices;
 import org.gradle.internal.service.ServiceRegistry;
 import org.gradle.internal.service.ServiceRegistryBuilder;
