+++ subprojects/logging/src/main/java/org/gradle/internal/logging/sink/ConsoleState.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.nativeintegration;
+package org.gradle.internal.logging.sink;
 
 import org.gradle.internal.nativeintegration.console.ConsoleDetector;
 import org.gradle.internal.nativeintegration.console.ConsoleMetaData;
